added search feature for sections page
