+++ frontend/types/classTypes.ts
@@ -8,24 +8,3 @@ export interface Course {
   createdAt?: Date;
   updatedAt?: Date;
 }
-
-interface Schedule {
-  days: string[];
-  time: string;
-}
-
-export interface Section {
-  courseCode: string;
-  section: string;
-  semester: string;
-  instructor: string;
-  schedule: Schedule;
-  location?: string;
-  capacity?: number;
-  enrolledStudents?: string[];
-  enrollmentStartDate?: Date;
-  enrollmentEndDate?: Date;
-  dropDeadline?: Date;
-  createdAt?: Date;
-  updatedAt?: Date;
-}
