+++ frontend/types/sectionTypes.ts
@@ -1,0 +1,22 @@
+interface Schedule {
+  days: string[];
+  time: string;
+}
+
+export interface Section {
+  courseCode: string;
+  section: string;
+  semester: string;
+  instructor: string;
+  schedule: Schedule;
+  location?: string;
+  capacity?: number;
+  enrolledStudents?: string[];
+  enrollmentStartDate?: Date;
+  enrollmentEndDate?: Date;
+  dropDeadline?: Date;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
+
+export const Seasons: string[] = ["Spring", "Summer", "Fall"]
